доработка проекта
