+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -138,4 +138,15 @@ struct QuizQuestion {
     let text: String
     let correctAnswer: Bool
 }
+struct QuizStepViewModel {
+    
+  let image: UIImage
+  let question: String
+  let questionNumber: String
+}
+struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
