--- MovieQuiz/Presentation/QuizResultsViewModel.swift
@@ -1,13 +1,0 @@
-//
-//  File.swift
-//  MovieQuiz
-//
-//  Created by Георгий Ксенодохов on 05.07.2023.
-//
-
-import Foundation
-struct QuizResultsViewModel {
-  let title: String
-  let text: String
-  let buttonText: String
-}
