--- MovieQuiz/Presentation/QuizStepViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  File.swift
-//  MovieQuiz
-//
-//  Created by Георгий Ксенодохов on 05.07.2023.
-//
-
-import UIKit
-struct QuizStepViewModel {
-    
-  let image: UIImage
-  let question: String
-  let questionNumber: String
-}
